Update README.md
This is my internship task @CodSoft.I was asked to create a basic calculator using CSS, HTML, and JavaScript, and also to implement an interactive interface with buttons for addition, subtraction, multiplication, and division operations. This has been an amazing experience for me as I have learned a lot through this weeks-long journey by completing these simple tasks in web development.